Add files via upload
I opted to use the Match function in Pandas, this match function requires the installation of html5lib in order to work. I left all the other print functions in as I see you can either retrieve the table numbers or the lengths. I used this to extract data from Wiki for our South African Rugby team. I committed the file to a CSV file.